Get a data generating function for simulating a trial design

Generates a data generating function that give p-values from normally distributed or binary data for use in the sim_trial() function.

Usage

get_trial_data_gen(
  corr_endpoints,
  effect_sizes,
  binary_endpoints = character(0),
  second_stage = FALSE
)

Arguments

corr_endpoints

correlation matrix specifying correlation between the endpoints

effect_sizes

data frame specifying the effect in each arm and group. See also details

binary_endpoints

which treatment endpoints are binary endpoints (all control endpoints referenced by those are also automatically treated as binary)

second_stage

bolean, specify if the data is generated for the first or second stage. If for the second stage, the trial_adapt_n function must have been used on the trial_design() object being used.

Value

A function taking two arguments, see details

Details

This gives back a function, which takes as its arguments a number n and a trial_design object and returns a matrix of n * "number of hypotheses" p-values. Each of those p-values is generated by drawing values for each patient as specified by the design object. For endpoint specified as continuous, the values are drawn from a multivariate normal distribution, with the correlation between endpoints given by corr_endpoints and the effect size by the group effect size from effect_sizes. For binary endpoints, the same data is generated (with the same correlation being applied), but instead of a mean, effect_sizes specifies the percentage of responders, which is generated by using an approrpiate cutoff for the continuous data. The $p$-values generated are from a $t$-test for continous data and from a two-proportion $Z$-test for binary data.

The argument effect_sizes should be given in the same format as the n_table argument for the trial_design(), but instead of the column n, there should be a column named for each endpoint, giving the effect size (either as a normal mean or a response rate, for continous or binary data) for this specific group and arm.

Examples

alpha_spending_f <- function(x, t) 2 - 2 * pnorm(qnorm(1 - x / 2) / sqrt(t))

design <- trial_design(
  endpoints = 2,
  n_table = data.frame(arm = c("control", "arm1"), n = c(10, 10)),
  weights = c(0.5, 0.5),
  test_m = rbind(c(0, 1), c(1, 0)),
  alpha = 0.025,
  alpha_spending_f = alpha_spending_f,
  t = 1 / 2,
  names_arms = "arm1",
  names_endpoints = c("cont", "bin")
)

data_gen <- get_trial_data_gen(
 corr_endpoints = rbind(c(1, 0.5), c(0.5, 1)),
 effect_sizes = data.frame(arm = c("control", "arm1"), cont = c(0, 0.4), bin
 = c(0.2, 0.5)),
 binary_endpoints = "bin"
)

data_gen(10, design)
#>             [,1]        [,2]
#>  [1,] 0.38448194 0.009538161
#>  [2,] 0.45049425 0.974519032
#>  [3,] 0.01997447 0.025480968
#>  [4,] 0.16210244 0.036819135
#>  [5,] 0.29690005 0.500000000
#>  [6,] 0.27038509 0.180655214
#>  [7,] 0.44508303 0.060667625
#>  [8,] 0.00606176 0.012309381
#>  [9,] 0.47329436 0.180655214
#> [10,] 0.05887742 0.001707396